# core-common-modelhub-interfaces/src/main/java/com/firefly/core/modelhub/interfaces/dtos/VirtualEntitySchemaDto.java
/*
 * Copyright 2025 Firefly Software Foundation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */


package com.firefly.core.modelhub.interfaces.dtos;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;
import io.swagger.v3.oas.annotations.media.Schema;

import java.util.List;
import java.util.UUID;

/**
 * DTO for virtual entity schema operations.
 * This represents the complete schema of a virtual entity, including all its fields.
 */
@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
@Schema(description = "Virtual entity schema definition")
public class VirtualEntitySchemaDto {

    @Schema(description = "Entity information")
    private VirtualEntityDto entity;

    @Schema(description = "Entity fields")
    private List<VirtualEntityFieldDto> fields;
}