# core-common-modelhub-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-modelhub
  description: ModelHub Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/modelhub
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/entities/{id}:
    get:
      tags:
        - Virtual Entities
      summary: Get a virtual entity by ID
      description: Returns a virtual entity definition by its ID
      operationId: getEntityById
      parameters:
        - name: id
          in: path
          description: ID of the entity to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VirtualEntityDto'
        '404':
          description: Entity not found
    put:
      tags:
        - Virtual Entities
      summary: Update a virtual entity
      description: Updates an existing virtual entity definition
      operationId: updateEntity
      parameters:
        - name: id
          in: path
          description: ID of the entity to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VirtualEntityDto'
        required: true
      responses:
        '200':
          description: Entity updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VirtualEntityDto'
        '400':
          description: Invalid input
        '404':
          description: Entity not found
    delete:
      tags:
        - Virtual Entities
      summary: Delete a virtual entity
      description: Deletes a virtual entity definition and all its fields and records
      operationId: deleteEntity
      parameters:
        - name: id
          in: path
          description: ID of the entity to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Entity deleted successfully
        '404':
          description: Entity not found
...
  /api/v1/entities:
    get:
      tags:
        - Virtual Entities
      summary: Get all virtual entities
      description: Returns a list of all virtual entity definitions
      operationId: getAllEntities
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VirtualEntityDto'
    post:
      tags:
        - Virtual Entities
      summary: Create a new virtual entity
      description: Creates a new virtual entity definition
      operationId: createEntity
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VirtualEntityDto'
        required: true
      responses:
        '201':
          description: Entity created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VirtualEntityDto'
        '400':
          description: Invalid input
...
  /api/v1/fields/entity/{entityId}:
    get:
      tags:
        - Virtual Entity Fields
      summary: Get all fields for an entity
      description: Returns a list of all fields defined for a specific virtual entity
      operationId: getFieldsByEntityId
      parameters:
        - name: entityId
          in: path
          description: ID of the entity to retrieve fields for
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VirtualEntityFieldDto'
        '404':
          description: Entity not found
  /api/v1/entities/{name}/schema:
    get:
      tags:
        - Virtual Entities
      summary: Get the schema for a virtual entity
      description: Returns the complete schema (entity definition and fields) for a virtual entity
      operationId: getEntitySchema
      parameters:
        - name: name
          in: path
          description: Name of the entity to retrieve the schema for
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VirtualEntitySchemaDto'
        '404':
          description: Entity not found
  /api/v1/entities/name/{name}:
    get:
      tags:
        - Virtual Entities
      summary: Get a virtual entity by name
      description: Returns a virtual entity definition by its name
      operationId: getEntityByName
      parameters:
        - name: name
          in: path
          description: Name of the entity to retrieve
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VirtualEntityDto'
        '404':
          description: Entity not found
components:
  schemas:
    VirtualEntityFieldDto:
      required:
        - entityId
        - fieldKey
        - fieldType
      type: object
      properties:
        id:
          type: string
          description: Unique identifier
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        entityId:
          type: string
          description: Parent entity ID
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        fieldKey:
          type: string
          description: Field key (used in code/API)
          example: firstName
        fieldLabel:
          type: string
          description: Field label (displayed to users)
          example: First Name
        fieldType:
          type: string
          description: Field data type
          example: string
        required:
          type: boolean
          description: Whether the field is required
          example: true
        options:
          type: object
          additionalProperties:
            type: object
            description: Additional field options (validation, UI hints, etc.)
          description: Additional field options (validation, UI hints, etc.)
        orderIndex:
          type: integer
          description: Display order index
          format: int32
          example: 1
        createdBy:
          type: string
          description: Created by user
          example: admin
        createdAt:
          type: string
          description: Creation timestamp
          format: date-time
        updatedBy:
          type: string
          description: Last updated by user
          example: admin
        updatedAt:
          type: string
          description: Last update timestamp
          format: date-time
      description: Field to create
    VirtualEntityDto:
      required:
        - active
        - name
      type: object
      properties:
        id:
          type: string
          description: Unique identifier
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        name:
          type: string
          description: Entity name
          example: customer
        description:
          type: string
          description: Entity description
          example: Customer entity for storing customer information
        version:
          type: integer
          description: Entity version
          format: int32
          example: 1
        active:
          type: boolean
          description: Whether the entity is active
          example: true
        createdBy:
          type: string
          description: Created by user
          example: admin
        createdAt:
          type: string
          description: Creation timestamp
          format: date-time
        updatedBy:
          type: string
          description: Last updated by user
          example: admin
        updatedAt:
          type: string
          description: Last update timestamp
          format: date-time
      description: Entity to create
...
    VirtualEntitySchemaDto:
      type: object
      properties:
        entity:
          $ref: '#/components/schemas/VirtualEntityDto'
        fields:
          type: array
          description: Entity fields
          items:
            $ref: '#/components/schemas/VirtualEntityFieldDto'
      description: Virtual entity schema definition